Refactor #488 - rename `core:data-format` to `core:component:displayformat`
