+++ core/component/displayformat/src/main/AndroidManifest.xml
