+++ core/component/displayformat/src/main/kotlin/com/nlab/reminder/core/component/displayformat/ui/DayOfWeekExt.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.data.model.ui
+package com.nlab.reminder.core.component.displayformat.ui
 
 import androidx.annotation.StringRes
 import com.nlab.reminder.core.translation.StringIds
