+++ settings.gradle.kts
@@ -62,6 +62,7 @@ include(
     ":core:androidx:transition",
     ":core:annotation",
     ":core:component:currenttime",
+    ":core:component:displayformat",
     ":core:component:tag",
     ":core:component:tag-test",
     ":core:component:usermessage",
@@ -70,7 +71,6 @@ include(
     ":core:data",
     ":core:data-di",
     ":core:data-ext",
-    ":core:data-format",
     ":core:data-impl",
     ":core:data-platform-impl",
     ":core:data-test",
